+++ Source/Project_LLL/Private/GAS/Ability/Player/LLL_PlayerGameplayAbilityBase.cpp
@@ -3,3 +3,17 @@
 
 #include "GAS/Ability/Player/LLL_PlayerGameplayAbilityBase.h"
 
+#include "Entity/Character/Player/LLL_PlayerBase.h"
+
+bool ULLL_PlayerGameplayAbilityBase::CanActivateAbility(const FGameplayAbilitySpecHandle Handle, const FGameplayAbilityActorInfo* ActorInfo, const FGameplayTagContainer* SourceTags, const FGameplayTagContainer* TargetTags, FGameplayTagContainer* OptionalRelevantTags) const
+{
+	bool Result = Super::CanActivateAbility(Handle, ActorInfo, SourceTags, TargetTags, OptionalRelevantTags);
+
+	const ILLL_PlayerDependencyInterface* PlayerActor = Cast<ILLL_PlayerDependencyInterface>(CurrentActorInfo->AvatarActor);
+	if(!PlayerActor)
+	{
+		return false;
+	}
+	
+	return Result;
+}
