+++ Source/Project_LLL/Private/Interface/LLL_PlayerDependencyInterface.cpp
@@ -1,0 +1,7 @@
+﻿// Fill out your copyright notice in the Description page of Project Settings.
+
+
+#include "Interface/LLL_PlayerDependencyInterface.h"
+
+
+// Add default functionality here for any ILLL_PlayerDependencyInterface functions that are not pure virtual.
