+++ Source/Project_LLL/Public/GAS/Ability/Player/LLL_PlayerGameplayAbilityBase.h
@@ -13,5 +13,10 @@ UCLASS()
 class PROJECT_LLL_API ULLL_PlayerGameplayAbilityBase : public ULLL_CharacterGameplayAbilityBase
 {
 	GENERATED_BODY()
+
+public:
+
+protected:
+	virtual bool CanActivateAbility(const FGameplayAbilitySpecHandle Handle, const FGameplayAbilityActorInfo* ActorInfo, const FGameplayTagContainer* SourceTags, const FGameplayTagContainer* TargetTags, FGameplayTagContainer* OptionalRelevantTags) const override;
 	
 };
