+++ Source/Project_LLL/Public/Interface/LLL_PlayerDependencyInterface.h
@@ -1,0 +1,25 @@
+﻿// Fill out your copyright notice in the Description page of Project Settings.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "UObject/Interface.h"
+#include "LLL_PlayerDependencyInterface.generated.h"
+
+// This class does not need to be modified.
+UINTERFACE()
+class ULLL_PlayerDependencyInterface : public UInterface
+{
+	GENERATED_BODY()
+};
+
+/**
+ * 
+ */
+class PROJECT_LLL_API ILLL_PlayerDependencyInterface
+{
+	GENERATED_BODY()
+
+	// Add interface functions to this class. This is the class that will be inherited to implement this interface.
+public:
+};
